Replaces the mandatory traits
In place of the mandatory traits we had, we now omit the command and event payloads in the case of where we are requesting the next events and replying the next event respectively. Serde has been leveraged to omit serialising an option's Some value when it is positioned at the end of a struct. Similarly, deserialisation tolerates the lack of these bytes resulting in a None type accordingly.