Avoid the varint
Due to encoding the data packet header as a u32, Postcard will employ its varint approach where, if the 16th bit of a frame counter is set, another byte will be expected to follow. This is undesirable as it will exceed the MIN_PACKET_SIZE when we use server discovery. We solve this by consuming some bits I had previously reserved.