Make provision for digital signing verification
Digital signatures such as those used by ed25519 can now be conveyed as the first n bytes of an update. If a signature is provided then this is also conveyed on the prepare-update request. The type and length of the signature are an application concern, but ed25519 is considered.

The reason we convey the signature up front is so that a digest can be formed by the server as bytes are received, rather than making another pass over the bytes to form a signature.